+++ GameEngine/src/main/java/com/bpham/gameengine/Port/GameEngine.java
@@ -1,7 +1,7 @@
-package com.bpham.gameengine.Port;
+package com.bpham.gameengine.port;
 
-import com.bpham.gameengine.Model.Monster;
-import com.bpham.gameengine.Model.MonsterQueueObservable;
+import com.bpham.gameengine.domain.Monster;
+import com.bpham.gameengine.domain.MonsterQueueObservable;
 
 import java.util.List;
