+++ GameEngine/src/main/java/com/bpham/gameengine/Port/MonsterDetailRepository.java
@@ -1,6 +1,5 @@
-package com.bpham.gameengine.Port;
+package com.bpham.gameengine.port;
 
-import java.util.Hashtable;
 import java.util.List;
 
 /**
