+++ GameEngine/src/main/java/com/bpham/gameengine/Port/RandomEncounterManager.java
@@ -1,4 +1,4 @@
-package com.bpham.gameengine.Port;
+package com.bpham.gameengine.port;
 
 /**
  * Created by Bryant on 8/17/2014.
