+++ GameEngine/src/test/java/com/bpham/gameengine/GameEngine/RandomPokemonFactoryTest.java
@@ -1,8 +1,8 @@
-package com.bpham.gameengine.GameEngine;
+package com.bpham.gameengine.gameengine;
 
-import com.bpham.gameengine.Model.Monster;
-import com.bpham.gameengine.Port.MonsterDetailRepository;
-import com.bpham.gameengine.Port.RandomMonsterFactory;
+import com.bpham.gameengine.domain.Monster;
+import com.bpham.gameengine.port.MonsterDetailRepository;
+import com.bpham.gameengine.port.RandomMonsterFactory;
 
 import org.junit.After;
 import org.junit.Before;
@@ -14,7 +14,6 @@ import org.mockito.MockitoAnnotations;
 import static org.hamcrest.CoreMatchers.*;
 
 import java.util.ArrayList;
-import java.util.Hashtable;
 import java.util.List;
 import java.util.Random;
