+++ app/src/main/java/com/pokebro/Repository/CaughtPokemonTable.java
@@ -1,4 +1,4 @@
-package com.pokebro.Repository;
+package com.pokebro.repository;
 
 import android.provider.BaseColumns;
