+++ app/src/main/java/com/pokebro/android/activity/MonsterQueueFragment.java
@@ -9,13 +9,12 @@ import android.widget.Button;
 import android.widget.ListView;
 import android.widget.Toast;
 
-import com.bpham.gameengine.Model.Monster;
-import com.bpham.gameengine.Model.MonsterQueueObservable;
-import com.bpham.gameengine.Port.GameEngine;
-import com.pokebro.PokebroApplication;
-import com.pokebro.Utility.PokemonQueueArrayAdapter;
+import com.bpham.gameengine.domain.Monster;
+import com.bpham.gameengine.domain.MonsterQueueObservable;
+import com.bpham.gameengine.port.GameEngine;
+import com.pokebro.support.PokemonQueueArrayAdapter;
 import com.pokebro.R;
-import com.pokebro.Utility.SwipeDismissListViewTouchListener;
+import com.pokebro.support.SwipeDismissListViewTouchListener;
 import com.pokebro.android.base.BaseFragment;
 
 import java.util.Observable;
