+++ app/src/main/java/com/pokebro/support/DrawerItemArrayAdapter.java
@@ -1,4 +1,4 @@
-package com.pokebro.Utility;
+package com.pokebro.support;
 
 import android.app.Activity;
 import android.content.Context;
@@ -7,10 +7,9 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.ImageView;
-import android.widget.ListView;
 import android.widget.TextView;
 
-import com.pokebro.Model.DrawerItem;
+import com.pokebro.domain.DrawerItem;
 import com.pokebro.R;
 
 import java.util.List;
